# specs/001-phase2-frontend/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Todo App API (Phase 2)
  version: 1.0.0
  description: API for managing todos and user authentication.

servers:
  - url: /api
    description: Base URL for all API requests.

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        name: { type: string }

    Task:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: string }
        title: { type: string }
        description: { type: string }
        completed: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    TaskInput:
      type: object
      required: [title]
      properties:
        title: { type: string }
        description: { type: string }

paths:
  /tasks:
    get:
      summary: List all tasks for the authenticated user.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: A list of tasks.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Task' }
    post:
      summary: Create a new task.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskInput' }
      responses:
        '201':
          description: The created task.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }

  /tasks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    put:
      summary: Update an existing task.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskInput' }
      responses:
        '200':
          description: The updated task.
    delete:
      summary: Delete a task.
      security: [{ bearerAuth: [] }]
      responses:
        '204':
          description: Successfully deleted.

  /tasks/{id}/complete:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    patch:
      summary: Toggle task completion status.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: The updated task with toggled completion.